makefile.vc: remove Unix commands in lib_dist target (#3986)
makefile.vc was calling rm, cp and zip inside of the windows makefile. This was causing an error when I was building the windows build on my local machine. I have updated the section to check if the $(DISTDIR) exists and if not make the directory and then copy the gdal build artefacts into the $(DISTDIR)